+++ src/referencedata-supply-partner/supply-partner-association-data-builder.spec.js
@@ -72,7 +72,7 @@
         }
 
         function build() {
-            return new SupplyPartnerAssociation(this.buildJson());
+            return this.buildJson();
         }
 
         function buildWithFacilitiesAndOrderables() {